Update tabs and payment methods component classnames to meet standards
